Add license
Summary
-------

This PR preserves the original GPL v3.0 license as is required.

The comment "# Stolen from: Aaron Kuehler" in the install.sh script
indicates that this was copied from the indiebrain `.files` repository
. The indiebrain `.files` repository is released under the GPL v3.0
license which permits such derived works, but requires that the
copyright and license notices must be preserved.

References
----------

- https://github.com/kuhy/.dotfiles/blob/fcc0ce891d3d92e46abd09f8ed1a73a2667153b7/install.sh#L3
- https://github.com/indiebrain/.files